+++ packages/serverless-nextjs-component/__tests__/next-aws-cloudfront.response.test.js
@@ -1,3 +1,4 @@
+const zlib = require("zlib");
 const create = require("../next-aws-cloudfront");
 
 describe("Response Tests", () => {
@@ -222,6 +223,47 @@ describe("Response Tests", () => {
     });
   });
 
+  it("hasHeader", () => {
+    const { res } = create({
+      request: {
+        path: "/",
+        headers: {}
+      }
+    });
+    res.setHeader("x-custom-1", "1");
+
+    expect(res.hasHeader("x-custom-1")).toBe(true);
+    expect(res.hasHeader("x-custom-2")).toBe(false);
+  });
+
+  it("case insensitive headers", () => {
+    const { res } = create({
+      request: {
+        path: "/",
+        headers: {}
+      }
+    });
+    res.setHeader("x-custom-1", "1");
+    res.setHeader("X-CUSTOM-2", "2");
+    res.setHeader("X-cUsToM-3", "3");
+
+    expect(res.getHeader("X-CUSTOM-1")).toEqual("1");
+    expect(res.getHeader("x-custom-2")).toEqual("2");
+    expect(res.getHeader("x-CuStOm-3")).toEqual("3");
+
+    expect(res.getHeaders()).toEqual({
+      "x-custom-1": "1",
+      "x-custom-2": "2",
+      "x-custom-3": "3"
+    });
+
+    res.removeHeader("X-CUSTOM-1");
+    res.removeHeader("x-custom-2");
+    res.removeHeader("x-CUSTom-3");
+
+    expect(res.getHeaders()).toEqual({});
+  });
+
   it(`res.write('ok')`, () => {
     const { res, responsePromise } = create({
       request: {
@@ -250,6 +292,36 @@ describe("Response Tests", () => {
 
     return responsePromise.then(response => {
       expect(response.body).toEqual("b2s=");
+    });
+  });
+
+  it(`gzips`, () => {
+    const gzipSpy = jest.spyOn(zlib, "gzipSync");
+    gzipSpy.mockReturnValueOnce(Buffer.from("ok-gzipped"));
+
+    const { res, responsePromise } = create({
+      request: {
+        path: "/",
+        headers: {
+          "accept-encoding": [
+            {
+              key: "Accept-Encoding",
+              value: "gzip"
+            }
+          ]
+        }
+      }
+    });
+
+    res.end("ok");
+
+    gzipSpy.mockRestore();
+
+    return responsePromise.then(response => {
+      expect(response.headers["content-encoding"]).toEqual([
+        { key: "Content-Encoding", value: "gzip" }
+      ]);
+      expect(response.body).toEqual("b2stZ3ppcHBlZA=="); // "ok-gzipped" base64 encoded
     });
   });
 });