+++ packages/next-aws-lambda/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.8](https://github.com/danielcondemarin/serverless-nextjs-plugin/compare/next-aws-lambda@2.3.7...next-aws-lambda@2.3.8) (2020-03-29)
+
+**Note:** Version bump only for package next-aws-lambda
+
+
+
+
+
+## [2.3.7](https://github.com/danielcondemarin/serverless-nextjs-plugin/compare/next-aws-lambda@2.3.6...next-aws-lambda@2.3.7) (2020-03-29)
+
+**Note:** Version bump only for package next-aws-lambda
+
+
+
+
+
+## [2.3.6](https://github.com/danielcondemarin/serverless-nextjs-plugin/compare/next-aws-lambda@2.3.5...next-aws-lambda@2.3.6) (2020-03-22)
+
+
+### Bug Fixes
+
+* handle empty req url ([#315](https://github.com/danielcondemarin/serverless-nextjs-plugin/issues/315)) ([ad7bec1](https://github.com/danielcondemarin/serverless-nextjs-plugin/commit/ad7bec1827ad3b6074c6f1a085a57a2d906334ba))
+
+
+
+
+
 ## 2.3.5 (2020-02-29)